Services/SensorEvaluationService/requirements.txt
rdflib
rdflib-jsonld
kafka-python
pandas
numpy
pyprog
requests
SPARQLWrapper